+++ src/renderer/features/multisig-operations/components/EmptyState/EmptyOperations.test.tsx
