+++ src/renderer/features/multisig-operations/components/EmptyState/EmptyOperations.tsx
