+++ src/renderer/features/multisig-operations/components/modals/RejectTx.tsx
