+++ src/renderer/features/multisig-operations/components/modals/SignatorySelectModal.tsx
