+++ src/renderer/features/multisig-operations/index.ts
@@ -1,14 +1,6 @@
-import { logTitleSlot, operationDetailsSlot, operationTitleSlot } from './components/Operation';
-import { operationsModel } from './model/list';
-
-export const multisigOperationsFeature = {
-  model: {
-    operations: operationsModel,
-  },
-  views: {},
-  slots: {
-    operationDetails: operationDetailsSlot,
-    operationTitle: operationTitleSlot,
-    logTitle: logTitleSlot,
-  },
-};
+export { logTitleSlot } from './components/LogModal';
+export { operationTitleSlot } from './components/Operation';
+export { operationDetailsSlot } from './components/OperationFullInfo';
+export { Operations } from './components/Operations';
+export { operationsModel } from './model/list';
+export { SignatorySelectModal } from './components/modals/SignatorySelectModal';
