+++ src/renderer/features/multisig-operations/model/flexible-shell-model.ts
