+++ src/renderer/features/multisig-operations/model/reject-model.ts
