--- src/renderer/pages/Operations/common/constants.ts
@@ -1,4 +1,0 @@
-export const InteractionStyle =
-  'rounded hover:bg-action-background-hover hover:text-text-primary cursor-pointer py-[3px] px-2 -mr-2';
-
-export const DescriptionBlockStyle = 'rounded bg-block-background px-3 py-2 flex flex-col gap-x-0.5 mb-2';
