--- src/renderer/pages/Operations/components/Operation.tsx
@@ -1,32 +1,0 @@
-import { type FlexibleMultisigTransactionDS, type MultisigTransactionDS } from '@/shared/api/storage';
-import { type FlexibleMultisigAccount, type MultisigAccount } from '@/shared/core';
-import { Slot } from '@/shared/di';
-import { Accordion } from '@/shared/ui';
-import { OperationTitleDate, OperationTitleStatus } from '@/entities/operations';
-import { multisigOperationsFeature } from '@/features/multisig-operations';
-
-import { OperationFullInfo } from './OperationFullInfo';
-
-type Props = {
-  tx: MultisigTransactionDS | FlexibleMultisigTransactionDS;
-  account: MultisigAccount | FlexibleMultisigAccount | null;
-};
-
-const Operation = ({ tx, account }: Props) => {
-  return (
-    <Accordion className="rounded bg-block-background-default transition-shadow hover:shadow-card-shadow focus-visible:shadow-card-shadow">
-      <Accordion.Button buttonClass="px-2" iconWrapper="px-1.5">
-        <div className="flex h-[52px] w-full items-center gap-x-4 overflow-hidden">
-          <OperationTitleDate operation={tx} />
-          <Slot id={multisigOperationsFeature.slots.operationTitle} props={{ operation: tx }} />
-          <OperationTitleStatus operation={tx} />
-        </div>
-      </Accordion.Button>
-      <Accordion.Content className="border-t border-divider">
-        <OperationFullInfo tx={tx} account={account} />
-      </Accordion.Content>
-    </Accordion>
-  );
-};
-
-export default Operation;
