+++ src/renderer/pages/Operations/index.ts
@@ -1,2 +1,1 @@
 export { Operations } from './Operations';
-export { SignatorySelectModal } from './components/modals/SignatorySelectModal';
